% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bootstrap_precision}
\alias{bootstrap_precision}
\title{Bootstrap-based Relative Precision Analysis}
\usage{
bootstrap_precision(
  y,
  epsilon_rel = NULL,
  epsilon = NULL,
  n_boot = 1000,
  conf = 0.95,
  seed = 123,
  N_pop = NULL
)
}
\arguments{
\item{y}{Numeric vector of observations.}

\item{epsilon_rel}{Numeric. Desired relative precision threshold (e.g. 0.2).}

\item{epsilon}{Numeric. Desired absolute precision threshold (e.g 1)}

\item{n_boot}{Number of bootstrap replicates (default = 1000).}

\item{conf}{Confidence level (default = 0.95).}

\item{seed}{Random seed (default = 123).}

\item{N_pop}{Finite population size (default = NULL).}
}
\value{
List with table of results and first sample size reaching desired precision.
}
\description{
Estimates sampling distribution of the mean using bootstrap, compute confidence intervals,
and track relative precision across increasing sample sizes. It calculates the precision for both an
infinite population size or from a complete sample (finite population size).
}
\examples{
y <- rpois(30, 5)
bootstrap_precision(y, epsilon_rel = 0.2, n_boot = 200)
bootstrap_precision(y, epsilon_rel = 0.2, n_boot = 200, N_pop=100)
}
