% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{finite_population_correction}
\alias{finite_population_correction}
\title{Finite Population Correction for Sample Size}
\usage{
finite_population_correction(n0, N)
}
\arguments{
\item{n0}{Numeric. Initial sample size assuming an infinite population.}

\item{N}{Numeric. Total population size.}
}
\value{
Adjusted sample size (numeric).
}
\description{
Adjusts an initial sample size calculation (n0) for a finite population.
}
\examples{
n0 <- 50     # initial sample size
N  <- 200    # total population size
finite_population_correction(n0, N)
}
