% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picinguaba.R
\docType{data}
\name{picinguaba}
\alias{picinguaba}
\title{picinguaba Dataset}
\format{
A data frame with 180 rows and 4 variables:
\describe{
  \item{Lumbrineris}{density of the polychaete species *Lumbrineris curtolobata* }
  \item{Amphitalamus}{density of the gastropod species *Amphitalamus vallei*}
  \item{Hemipholis}{density of the ophiuroid species *Hemipholis elongata*}
  \item{Magelona}{density of the polychaete species *Magelona pappillicornis*}
}
}
\source{
Paiva, P.C. 2001. Spatial and Temporal Variation of a Nearshore Benthic Community in Southern Brazil: Implications for the Design of Monitoring Programs,
Estuarine, Coastal and Shelf Science, 52(4):423-433. https://doi.org/10.1006/ecss.2001.0763.
}
\usage{
picinguaba
}
\description{
The `picinguaba` dataset contains data on counts of benthic species collected in shallow waters in Picinguaba bay Ubatuba city, São Paulo, Brazil.
}
\details{
The dataset includes:
Density for  0.008 square meters of benthic species from 10 m depth nearshore sediments.
}
\examples{
data("picinguaba")
head(picinguaba)
}
\keyword{datasets}
