% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_distributionsC}
\alias{plot_distributionsC}
\title{Histogram with Fitted Distributions (Continuous)}
\usage{
plot_distributionsC(y)
}
\arguments{
\item{y}{Numeric vector of positive values.}
}
\value{
A ggplot object.
}
\description{
Plots a histogram of continuous or positive data and overlays
Normal, Student-t,  Gamma (MLE), and Lognormal (MLE) distributions.
}
\examples{
y <- rgamma(100, shape = 2, rate = 0.5)
plot_distributionsC(y)
}
