% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{relative_sample_zip}
\alias{relative_sample_zip}
\title{Sample Size Calculation for Relative Error with ZIP}
\usage{
relative_sample_zip(y, epsilon_rel, conf = 0.95, N = NULL)
}
\arguments{
\item{y}{Numeric vector of counts.}

\item{epsilon_rel}{Desired relative error (0.1)}

\item{conf}{Confidence level (default 0.95)}

\item{N}{Population size (optional, for finite population correction)}
}
\value{
Prints required sample sizes.
}
\description{
Computes sample sizes for both Discrete (Normal, Student-t, Poisson, and Negative Binomial) or
Continuous (Normal, Lognormal and Gamma) distributions.
}
\examples{
y <- rpois(100, 3); y[sample(1:100, 20)] <- 0
relative_sample_zip(y, 0.1)
}
