% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sample_size_beta}
\alias{sample_size_beta}
\title{Sample Size Calculation for a Beta Distribution Mean}
\usage{
sample_size_beta(
  x = NULL,
  alpha = NULL,
  beta = NULL,
  epsilon,
  conf = 0.95,
  N = NULL,
  relative = FALSE
)
}
\arguments{
\item{x}{Numeric vector of data in [0,1], or NULL if user provides alpha and beta.}

\item{alpha}{Numeric. Shape1 parameter of the Beta distribution (optional if x given).}

\item{beta}{Numeric. Shape2 parameter of the Beta distribution (optional if x given).}

\item{epsilon}{Numeric. Desired absolute precision (margin of error, >0).}

\item{conf}{Numeric. Confidence level (default = 0.95).}

\item{N}{Optional. Population size for finite population correction (default = NULL).}

\item{relative}{Logical. If TRUE, interprets E as relative error (default = FALSE).}
}
\value{
Required sample size (integer).
}
\description{
Computes the required sample size to estimate the mean of a Beta distribution
with a given absolute precision (margin of error) and confidence level.
Automatically transforms any 0/1 values in the data using Smithson & Verkuilen (2006).
}
\examples{
# Example with observed data containing 0
x <- c(0.2, 0.1, 0, 0, 0, 0.4)
sample_size_beta(x = x, epsilon = 0.05)

# Example with provided parameters
sample_size_beta(alpha = 2, beta = 5, epsilon = 0.05, conf = 0.95, N = 500)
}
