% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sargassum.R
\docType{data}
\name{sargassum}
\alias{sargassum}
\title{sargassum Dataset}
\format{
A data frame with <number of rows> rows and <number of columns> variables:
\describe{
  \item{Bra_Density}{Density of sargassum in Brandão Island}
  \item{Bra_Cover}{canopy cover of sargassum in Brandão Island}
  \item{Bra_PA}{Presence/absence(0/1) of sargassum in Brandão Island}
  \item{Buz_Density}{Density of sargassum in Buzios Island}
  \item{Buz_Cover}{canopy cover of sargassum in Buzios Island}
  \item{Buz_PA}{Presence/absence(0/1) of sargassum in Buzios Island}
  ...
}
}
\source{
Carneiro, I.M., Paiva, P.C., Bertocci, I. & Széchy, M.T.M. 2020. Reconciling vertical and horizontal variability in Sargassum populations
for improved environmental monitoring. Journal of  Applied Phycology 32:717–728.. https://doi.org/10.1007/s10811-019-01882-x
}
\usage{
sargassum
}
\description{
Samples of *Sargassum spp.* collected from subtidal rocky habitats on six islands in Ilha Grande Bay, Brazil.
The dataset includes measurements of canopy cover, density, and presence/absence of species for each island.
}
\examples{
data("sargassum")
head(sargassum)
}
\keyword{datasets}
